+++ pytest_lambda/fixtures.py
@@ -1,17 +1,32 @@
+from __future__ import annotations
+
 import inspect
-from typing import Union, Callable, Any, Iterable
+from typing import NoReturn, TYPE_CHECKING, Callable, Any, Iterable, Tuple, TypeVar
 
 from pytest_lambda.exceptions import DisabledFixtureError, NotImplementedFixtureError
 from pytest_lambda.impl import LambdaFixture
+
+if TYPE_CHECKING:
+    from _pytest.fixtures import _Scope
 
 __all__ = ['lambda_fixture', 'static_fixture', 'error_fixture',
            'disabled_fixture', 'not_implemented_fixture']
 
 
-def lambda_fixture(fixture_name_or_lambda: Union[str, Callable]=None,
-                   *other_fixture_names: Iterable[str],
-                   bind: bool = False, async_: bool = False,
-                   scope="function", params=None, autouse=False, ids=None, name=None):
+VT = TypeVar('VT')
+
+
+def lambda_fixture(
+    fixture_name_or_lambda: str | Callable[..., VT] | None = None,
+    *other_fixture_names: str,
+    bind: bool = False,
+    async_: bool = False,
+    scope: _Scope = 'function',
+    params: Iterable[object] | None = None,
+    autouse: bool = False,
+    ids: Iterable[None | str | float | int | bool] | Callable[[Any], object | None] | None = None,
+    name: str | None = None,
+) -> LambdaFixture[VT]:
     """Use a fixture name or lambda function to compactly declare a fixture
 
     Usage:
@@ -32,20 +47,32 @@ def lambda_fixture(fixture_name_or_lambda: Union[str, Callable]=None,
     :param async_: If True, the lambda will be wrapped in an async function; if the
         lambda evaluates to an awaitable value, it will be awaited.
 
+    :param scope:
+    :param params:
+    :param autouse:
+    :param ids:
+    :param name:
+        Options to pass to pytest.fixture()
+
     """
+    fixture_names_or_lambda: Tuple[str | Callable, ...] | str | Callable | None
+
     if other_fixture_names:
+        if fixture_name_or_lambda is None:
+            raise ValueError('If specified, all fixture names must be non-null.')
         fixture_names_or_lambda = (fixture_name_or_lambda,) + other_fixture_names
     else:
         fixture_names_or_lambda = fixture_name_or_lambda
 
     return LambdaFixture(
         fixture_names_or_lambda,
-        bind=bind, async_=async_,
+        bind=bind,
+        async_=async_,
         scope=scope, params=params, autouse=autouse, ids=ids, name=name,
     )
 
 
-def static_fixture(value: Any, **fixture_kwargs):
+def static_fixture(value: VT, **fixture_kwargs) -> LambdaFixture[VT]:
     """Compact method for defining a fixture that returns a static value
     """
     return lambda_fixture(lambda: value, **fixture_kwargs)
@@ -59,7 +86,7 @@ def raise_exception({args}):
 '''
 
 
-def error_fixture(error_fn: Callable, **fixture_kwargs):
+def error_fixture(error_fn: Callable, **fixture_kwargs) -> LambdaFixture[NoReturn]:
     """Fixture whose usage results in the raising of an exception
 
     Usage:
@@ -89,7 +116,7 @@ def error_fixture(error_fn: Callable, **fixture_kwargs):
     return lambda_fixture(raise_exception, **fixture_kwargs)
 
 
-def disabled_fixture(**fixture_kwargs):
+def disabled_fixture(**fixture_kwargs) -> LambdaFixture[NoReturn]:
     """Mark a fixture as disabled – using the fixture will raise an error
 
     This is useful when you know any usage of a fixture would be in error. When
@@ -116,7 +143,7 @@ def disabled_fixture(**fixture_kwargs):
     return error_fixture(build_disabled_fixture_error, **fixture_kwargs)
 
 
-def not_implemented_fixture(**fixture_kwargs):
+def not_implemented_fixture(**fixture_kwargs) -> LambdaFixture[NoReturn]:
     """Mark a fixture as abstract – requiring definition/override by the user
 
     This is useful when defining abstract base classes requiring implementation